+++ src/router/index.tsx
@@ -1,22 +1,11 @@
-import { Outlet, Route, createBrowserRouter, createRoutesFromChildren } from 'react-router-dom';
-import { styled } from 'styled-components';
+import { Route, createBrowserRouter, createRoutesFromChildren } from 'react-router-dom';
+import ErrorPage from '../components/ErrorPage';
 import IssueDetail from '../pages/IssueDetail';
 import Issues from '../pages/Issues';
-import ErrorPage from '../components/ErrorPage';
+import Layout from '../components/Layout';
 
 export const OWNER = 'facebook';
 export const REPO = 'react';
-
-const Layout = () => (
-  <Container>
-    <RepoHeader>
-      <RepoOwner>{OWNER}</RepoOwner>
-      <RepoName>/</RepoName>
-      <RepoName>{REPO}</RepoName>
-    </RepoHeader>
-    <Outlet />
-  </Container>
-);
 
 const router = createBrowserRouter(
   createRoutesFromChildren(
@@ -30,31 +19,3 @@ const router = createBrowserRouter(
 );
 
 export default router;
-
-const Container = styled.div`
-  background-color: #181a1f;
-  padding-right: 20px;
-  padding-left: 20px;
-  font-family: Arial, sans-serif;
-  height: 100vh;
-`;
-
-const RepoHeader = styled.div`
-  background-color: #20232a;
-  padding-left: 16px;
-  display: flex;
-  align-items: center;
-  font-family: Arial, sans-serif;
-  font-size: 20px;
-  height: 54.5px;
-`;
-
-const RepoOwner = styled.span`
-  color: #61dafb;
-  margin-right: 8px;
-`;
-
-const RepoName = styled.span`
-  color: #e1e4e8;
-  font-weight: bold;
-`;
